Refactor RecentlyPlayed component: extract logic into useSpotify composable and streamline data fetching
